-#710 changed any type to given interface
put interface in workpackages

install prietter
