+++ src/frontend/src/apis/work-packages.api.ts
@@ -4,10 +4,22 @@
  */
 
 import axios from '../utils/axios';
-import { WbsNumber, WorkPackage } from 'shared';
+import { WbsNumber, WorkPackage, WorkPackageStage } from 'shared';
 import { wbsPipe } from '../utils/pipes';
 import { apiUrls } from '../utils/urls';
 import { workPackageTransformer } from './transformers/work-packages.transformers';
+
+export interface CreateWorkPackageFormInputs {
+  name: string;
+  startDate: Date;
+  duration: number;
+  crId: string;
+  stage: WorkPackageStage | null;
+  wbsNum: string;
+  blockedBy: { wbsNum: string }[];
+  expectedActivities: { bulletId: number; detail: string }[];
+  deliverables: { bulletId: number; detail: string }[];
+}
 
 /**
  * Fetch all work packages.
@@ -34,7 +46,7 @@ export const getSingleWorkPackage = (wbsNum: WbsNumber) => {
  *
  * @param payload Payload containing all the necessary data to create a work package.
  */
-export const createSingleWorkPackage = (payload: any) => {
+export const createSingleWorkPackage = (payload: CreateWorkPackageFormInputs) => {
   return axios.post<{ message: string }>(apiUrls.workPackagesCreate(), {
     ...payload
   });
@@ -46,7 +58,7 @@ export const createSingleWorkPackage = (payload: any) => {
  * @param payload Object containing required key-value pairs for backend function to edit work package
  * @returns Promise that will resolve to either a success status code or a fail status code.
  */
-export const editWorkPackage = (payload: any) => {
+export const editWorkPackage = (payload: CreateWorkPackageFormInputs) => {
   return axios.post<{ message: string }>(apiUrls.workPackagesEdit(), {
     ...payload
   });
